requirements.txt
rich
google-generativeai
python-dotenv
